+++ zeeguu_dashboard/app/forms/create_cohort.py
@@ -1,20 +1,16 @@
 from flask_wtf import FlaskForm
 from wtforms import StringField, SubmitField
 from wtforms.validators import DataRequired
-
 from app.util.classroom import verify_invite_code_exists
 
 """
-This file contains all of the forms used in this system:
-- form for creating a new class
-- form for editing a class
-- form for logging in
+The create cohort form class file.
 """
-
 
 class CreateCohort(FlaskForm):
     """
-
+    This class extends from FlaskForm. It is used for the form when filling out
+    the information of a new class.
     """
     class_name = StringField('Class room name', validators=[DataRequired()])
     inv_code = StringField('Invite code', validators=[DataRequired()])
@@ -24,8 +20,10 @@ class CreateCohort(FlaskForm):
 
     def validate(self):
         """
-
-        :return:
+        This function validates the CreateCohort form.
+        It extends from the normal validation as we need to validate whether some
+        filled in data is already in use (the class invite code).
+        :return: Returns a boolean, indicating whether the form is properly filled out or not.
         """
         if not FlaskForm.validate(self):
             return False
@@ -35,22 +33,3 @@ class CreateCohort(FlaskForm):
             self.inv_code.errors = tuple(tmp)
             return False
         return True
-
-
-class EditCohort(FlaskForm):
-    """
-
-    """
-    class_name = StringField('Class room name', validators=[DataRequired()])
-    inv_code = StringField('Invite code')
-    max_students = StringField('Max students', validators=[DataRequired()])
-    submit = SubmitField('Create classroom')
-
-
-class CreateLogin(FlaskForm):
-    """
-
-    """
-    email = StringField('username', validators=[DataRequired()])
-    password = StringField('password', validators=[DataRequired()])
-    submit = SubmitField('Login')
