+++ zeeguu_dashboard/app/page_routes/errorpages.py
@@ -10,20 +10,18 @@ This file contains all of the error page routes.
 @app.errorhandler(401)
 @app.route("/401")
 def invalid_credentials():
-    print("called 401")
     """
-       Function for loading the 401 error page when page when logged in user and unautherised.
-       :return: Renders and returns an error page.
+    Function for loading the 401 error page when page when logged in user and unautherised.
+    :return: Renders and returns an error page.
     """
-    return render_template("errorpage.html", exception="401 UNAUTHORIZED")
+    return render_template("errorpage.html", exception="401 Unauthorized: You do not have access to this page.")
 
 
 @app.errorhandler(404)
 def page_not_found(e):
-    print("called 404")
     """
-       Function for loading the 404 error page when requested url does not exist.
-       :return: Renders and returns an error page.
+    Function for loading the 404 error page when requested url does not exist.
+    :return: Renders and returns an error page.
     """
     return render_template("errorpage.html", exception="404 Not Found: The requested URL was not found on the server.")
